+++ packages/frontend/src/lib/progress/TasksProgress.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 import type { Task } from '@shared/src/models/ITask';
 import ExpandableMessage from '/@/lib/ExpandableMessage.svelte';
-import TaskProgress from './TaskProgress.svelte';
+import TaskItem from './TaskItem.svelte';
 
 export let tasks: Task[] = [];
 </script>
@@ -9,7 +9,7 @@ export let tasks: Task[] = [];
 <ul class="space-y-2 text-gray-500 list-inside dark:text-gray-400">
   {#each tasks as task}
     <li class="flex flex-col rounded-md bg-charcoal-800 p-2">
-      <TaskProgress task="{task}" />
+      <TaskItem task="{task}" />
       <ExpandableMessage message="{task.error}" title="View Error" />
     </li>
   {/each}
